+++ api/machine/machine.proto
@@ -11,13 +11,6 @@ import "google/protobuf/any.proto";
 import "google/protobuf/empty.proto";
 import "google/protobuf/timestamp.proto";
 import "common/common.proto";
-import "storage/storage.proto";
-
-service MaintenanceService {
-  rpc ApplyConfiguration(ApplyConfigurationRequest)
-      returns(ApplyConfigurationResponse);
-  rpc Disks(google.protobuf.Empty) returns (storage.DisksResponse);
-}
 
 // The machine service definition.
 service MachineService {
