+++ internal/app/maintenance/server/server.go
@@ -9,11 +9,10 @@ import (
 	"fmt"
 	"log"
 
-	"github.com/golang/protobuf/ptypes/empty"
-	"github.com/talos-systems/go-blockdevice/blockdevice/util"
 	"google.golang.org/grpc"
 
 	"github.com/talos-systems/talos/internal/app/machined/pkg/runtime"
+	storaged "github.com/talos-systems/talos/internal/app/storaged"
 	"github.com/talos-systems/talos/pkg/machinery/api/machine"
 	"github.com/talos-systems/talos/pkg/machinery/api/storage"
 	"github.com/talos-systems/talos/pkg/machinery/config/configloader"
@@ -21,6 +20,8 @@ import (
 
 // Server implements machine.MaintenanceService.
 type Server struct {
+	machine.UnimplementedMachineServiceServer
+	storaged.Server
 	runtime runtime.Runtime
 	cfgCh   chan []byte
 	logger  *log.Logger
@@ -40,11 +41,11 @@ func New(r runtime.Runtime, logger *log.Logger, cfgCh chan []byte) *Server {
 func (s *Server) Register(obj *grpc.Server) {
 	s.server = obj
 
-	machine.RegisterMaintenanceServiceServer(obj, s)
 	storage.RegisterStorageServiceServer(obj, s)
+	machine.RegisterMachineServiceServer(obj, s)
 }
 
-// ApplyConfiguration implements machine.MaintenanceService.
+// ApplyConfiguration implements machine.MachineService.
 func (s *Server) ApplyConfiguration(ctx context.Context, in *machine.ApplyConfigurationRequest) (reply *machine.ApplyConfigurationResponse, err error) {
 	cfgProvider, err := configloader.NewFromBytes(in.GetData())
 	if err != nil {
@@ -62,30 +63,6 @@ func (s *Server) ApplyConfiguration(ctx context.Context, in *machine.ApplyConfig
 	}
 
 	s.cfgCh <- in.GetData()
-
-	return reply, nil
-}
-
-// Disks implements machine.MaintenanceService.
-func (s *Server) Disks(ctx context.Context, in *empty.Empty) (reply *storage.DisksResponse, err error) {
-	disks, err := util.GetDisks()
-	if err != nil {
-		return nil, err
-	}
-
-	diskList := make([]*storage.Disk, len(disks))
-
-	for i, disk := range disks {
-		diskList[i] = &storage.Disk{
-			DeviceName: disk.DeviceName,
-			Model:      disk.Model,
-			Size:       disk.Size,
-		}
-	}
-
-	reply = &storage.DisksResponse{
-		Disks: diskList,
-	}
 
 	return reply, nil
 }
